backend/requirements.txt
# server things 

fastapi[standard]
fastapi-cors
uvicorn[standard]
fastapi-mail
firebase-admin
pyjwt

# mqtt communitaion 
paho-mqtt
asyncio-mqtt

# testing 
pytest
pytest-asyncio
pytest-cov # test coverage

# db 
pydantic
pymongo
motor

# env vars 
python-decouple

# stats 
pandas
numpy

# hashing 
passlib[bcrypt]